pyproject.toml
[tool.poetry]
name = "scholar_verse"
version = "0.1.0"
description = "Autonomous Scientific Knowledge Graph Platform"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
google-adk = "^0.1.0"
fastapi = "^0.115.0"
uvicorn = "^0.34.0"
python-dotenv = "^1.0.0"
neo4j = "^5.14.0"
redis = "^5.0.1"
qdrant-client = "^1.6.0"
pydantic = "^2.4.2"
pyPDF2 = "^3.0.1"
requests = "^2.31.0"
google-cloud-aiplatform = "^1.36.0"
google-cloud-storage = "^2.12.0"
google-cloud-bigquery = "^3.13.0"
vertexai = "^0.0.1"
# Deep Search dependencies
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.3"
nltk = "^3.8.1"
regex = "^2023.10.3"
numpy = "^1.26.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.9.1"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
scholar_verse = "scholar_verse.main:main"
